requirement.txt
pandas==1.3.4
plotly==5.5.0
streamlit==1.3.0
Pillow==8.4.0

